sidebar for skill statistics
